Fix Windows build: missing <csignal>, uint64→uint32 cast
Add <csignal> to platform.h Windows block for sig_atomic_t.
Add explicit uint32_t cast in send_burst_tcp to fix MSVC warning.

+++ src/sender.cpp
@@ -67,7 +67,7 @@ static void send_burst_tcp(const Config& cfg) {
     stats.init();
 
     printf("Sending TCP burst: %.2f MB...\n", cfg.burst_bytes / 1e6);
-    if (tcp_send_len_prefixed(fd, buf, cfg.burst_bytes)) {
+    if (tcp_send_len_prefixed(fd, buf, (uint32_t)cfg.burst_bytes)) {
         stats.print(cfg.burst_bytes + 4, "TX");
     }
